progress.csv, progressplot.pdf update (>270h released now)
The plot is a bit optimistic: Some tasks counted as 'released' are not in fact
available to students, because their taskgroup is still alpha.
Also, difficulty 'high' should not be counted, because it is not relevant to the mainstream of students.
Counted that way, we are at only 193h released.